api/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
requests
psycopg2-binary
python-dotenv
gunicorn